Add more upload history code.
