+++ FERN/Actors/FileUploadActor.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Hopp, Dan on 7/30/24.
 //
+// 8/7/2024 May not need actor. FileUploadClass is currently inuse.
 
 import Foundation
 import CryptoKit
